requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --all-extras --resolver=backtracking pyproject.toml
#
attrs==22.2.0
    # via
    #   flake8-bugbear
    #   pytest
black==23.1.0
    # via cooklang (pyproject.toml)
click==8.1.3
    # via black
coverage[toml]==7.1.0
    # via pytest-cov
exceptiongroup==1.1.0
    # via pytest
flake8==6.0.0
    # via
    #   cooklang (pyproject.toml)
    #   flake8-bugbear
    #   flake8-builtins
    #   pep8-naming
flake8-bugbear==23.1.20
    # via cooklang (pyproject.toml)
flake8-builtins==2.1.0
    # via cooklang (pyproject.toml)
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.0.0
    # via cooklang (pyproject.toml)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.0
    # via
    #   black
    #   pytest
pathspec==0.11.0
    # via black
pep8-naming==0.13.3
    # via cooklang (pyproject.toml)
platformdirs==3.0.0
    # via black
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytest==7.2.1
    # via
    #   cooklang (pyproject.toml)
    #   pytest-cov
pytest-cov==4.0.0
    # via cooklang (pyproject.toml)
python-frontmatter==1.0.0
    # via cooklang (pyproject.toml)
pyyaml==6.0
    # via python-frontmatter
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via mypy
